change in text format
Style: a change to text format

there are some lines break in this file, some of them contain
the continuation of another sentence. these sentences/sub-sentences
should start in a new line for better look and good explanation.

a change to syntax for: ``make htmldocs`` and ``make pdfdocs`` codes

this can be done as well, while maintaing wrap up guide.